dataset_processor_api/pyproject.toml
[project]
name = "dataset-processor"
version = "0.1.0"
description = "Dataset Processor (Study Project)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "asgi-correlation-id>=4.3.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "fastapi[standard]>=0.128.0",
    "minio>=7.2.5",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.46",
    "structlog>=25.4.0",
]

[dependency-groups]
dev = [
    "celery-types>=0.24.0",
    "httpx>=0.28.1",
    "interrogate>=1.7.0",
    "ipython>=9.9.0",
    "mkdocs-material>=9.6.22",
    "mkdocstrings[python]>=0.30.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-xdist>=3.6.1",
    "ruff>=0.14.14",
    "taskipy>=1.14.1",
    "testcontainers[minio,postgres]>=4.14.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["src", "tests"]
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
    "B",
    "UP",
    "SIM",
    "C4",
    "A",
    "ARG",
    "PIE",
    "RET",
    "SLF",
    "TCH",
    "TID",
    "RUF",
    "D",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.per-file-ignores]
"src/api/routes/datasets.py" = ["TCH002"]
"tests/**/*.py" = ["D"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.13"
strict = true
warn_unreachable = true
show_error_codes = true
pretty = true
files = ["src", "tests"]
exclude = "(^migrations/)"
plugins = ["pydantic.mypy"]


[tool.taskipy.tasks]
format = "uv run ruff format"
format-check = "uv run ruff format --check"
lint = "uv run ruff check"
lint-fix = "uv run ruff check --fix"
typecheck = "uv run mypy"
docstring-check = "uv run interrogate -c pyproject.toml src"
docs-serve = "uv run mkdocs serve"
docs-build = "uv run mkdocs build --strict"
docs-check = "uv run task docs-build"
qa = "uv run task format-check && uv run task lint && uv run task typecheck && uv run task docstring-check"
check = "uv run task qa && uv run task docs-check"

test = "uv run pytest -m 'not e2e'"
test-fast = "uv run task test"
test-parallel = "uv run pytest -m 'not e2e' -n 2 --dist loadfile"
test-all = "uv run pytest"
test-all-parallel = "uv run pytest -n 2 --dist load"
test-e2e = "uv run pytest -m e2e"
verify = "uv run task check && uv run task test"
verify-all = "uv run task check && uv run task test-all"

stack-up = "docker compose -f ../docker-compose.yml up -d"
stack-rebuild = "docker compose -f ../docker-compose.yml up -d --build"
stack-down = "docker compose -f ../docker-compose.yml down"
stack-ps = "docker compose -f ../docker-compose.yml ps"
stack-logs = "docker compose -f ../docker-compose.yml logs -f api worker web"

db-upgrade = "docker compose -f ../docker-compose.yml run --rm api alembic upgrade head"
db-downgrade = "docker compose -f ../docker-compose.yml run --rm api alembic downgrade -1"
db-current = "docker compose -f ../docker-compose.yml run --rm api alembic current"
db-history = "docker compose -f ../docker-compose.yml run --rm api alembic history"
db-revision = "docker compose -f ../docker-compose.yml run --rm api alembic revision --autogenerate"
migrate = "uv run task db-upgrade"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "e2e: end-to-end tests with real broker/worker integration",
]
filterwarnings = [
    "ignore:The @wait_container_is_ready decorator is deprecated:DeprecationWarning:testcontainers\\.core\\.waiting_utils",
]

[tool.interrogate]
fail-under = 85
ignore-init-method = true
ignore-init-module = true
ignore-magic = true
ignore-semiprivate = true
ignore-private = true
ignore-property-decorators = true
exclude = ["tests", "migrations"]
